pyproject.toml
[project]
name = "max"
version = "0.1.0"
description = "."
authors = [{ name = "Patrick O'Reilly"}]
requires-python = ">=3.10"
dependencies = [
    # Core LangChain
    "langchain>=0.2.0",
    "langchain-ollama>=0.1.0",
    "langchain-community>=0.2.0",

    # Vector store
    "chromadb>=0.5.0",

    # UI
    "textual>=0.70.0",
    "rich>=13.0.0",

   " google-adk",

    "litellm",

    # CLI
    "typer>=0.12.0",

    # Syntax highlighting utilities
    "pygments>=2.17.0"
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "mypy",
    "pre-commit"
]

[project.scripts]
max = "max.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/max"]